+++ scripts/serve/serve-cli.js
@@ -1,0 +1,22 @@
+const commander = require('commander');
+const serve = require('./serve');
+
+const packageJson = require('../../package.json');
+const loadDefaultWebpackConfig = require('./loadDefaultWebpackConfig');
+
+
+// Parse process arguments
+commander
+  .version(packageJson.version)
+  .option('--config <path>', 'The webpack config to serve.', undefined)
+  .option('--port <n>', 'The port the server should listen on.', parseInt)
+  .option('-p, --production', 'Passes the -p flag to the webpack config')
+  .parse(process.argv);
+
+const port = commander.port || process.env.PORT;
+
+serve({
+  config: loadDefaultWebpackConfig(commander.config),
+  port,
+  production: commander.production,
+});