+++ CHANGELOG.md
@@ -3,8 +3,11 @@ Changelog
 
 Unreleased
 ----------
+### Major Version Bump. See https://github.com/cerner/terra-toolkit/blob/master/docs/TerraToolkitUpgradeBuid-v3.0.0.md.
+
 ### Added
 * Added nl and nl-BE to supported locales
+
 
 2.11.0 - (April 11, 2018)
 ----------